Move and cleanup NewSiteCreationService
