+++ WordPress/src/main/AndroidManifest.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <manifest package="org.wordpress.android"
           xmlns:android="http://schemas.android.com/apk/res/android"
           xmlns:tools="http://schemas.android.com/tools"
@@ -668,7 +669,7 @@
             android:label="Site Creation Service" />
 
         <service
-            android:name=".ui.sitecreation.NewSiteCreationService"
+            android:name=".ui.sitecreation.creation.NewSiteCreationService"
             android:exported="false"
             android:label="Site Creation Service" />
